Separate base class for crypto tests from hmac
It will be handy when there will be more cryptography to test. In
that case, the CryptoBase class with a CryptoApplet could be used.
